requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
anyio=3.6.2=pyhd8ed1ab_0
ca-certificates=2022.9.24=h033912b_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.0=py37hc55c11b_1
click=8.1.3=py37hf985489_0
cryptography=37.0.1=py37hf6deb26_0
fastapi=0.85.1=pyhd8ed1ab_0
h11=0.14.0=pyhd8ed1ab_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py37hf985489_0
libcxx=14.0.6=h9765a3e_0
libffi=3.3=hb1e8313_2
libsqlite=3.39.4=ha978bb4_0
libzlib=1.2.13=hfd90126_4
markupsafe=2.1.1=py37h69ee0a8_1
ncurses=6.3=hca72f7f_3
openssl=1.1.1q=hfd90126_1
peewee=3.15.3=py37h192cc77_0
pip=22.2.2=py37hecd8cb5_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.2=py37h8052db5_0
pyjwt=2.6.0=pyhd8ed1ab_0
pymysql=1.0.2=py37hecd8cb5_1
python=3.7.13=hdfd78df_1
python-dotenv=0.21.0=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
readline=8.1.2=hca72f7f_1
setuptools=63.4.1=py37hecd8cb5_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlite=3.39.3=h707629a_0
starlette=0.20.4=pyhd8ed1ab_0
tk=8.6.12=h5d9f67b_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
uvicorn=0.19.0=py37hf985489_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=hca72f7f_0
zipp=3.9.0=pyhd8ed1ab_0
zlib=1.2.13=hfd90126_4

peewee~=3.15.3
python-dotenv~=0.21.0
fastapi~=0.85.1
pydantic~=1.10.2
cryptography~=37.0.1
pyjwt~=2.6.0